[#27] refactor: domain 하위 디렉토리 생성 및 클래스 이동으로 인한 변경
