+++ src/main/java/com/toy/getyourfriday/component/ProductContainer.java
@@ -1,7 +1,7 @@
 package com.toy.getyourfriday.component;
 
-import com.toy.getyourfriday.domain.ModelUrl;
-import com.toy.getyourfriday.domain.Products;
+import com.toy.getyourfriday.domain.product.Products;
+import com.toy.getyourfriday.domain.scraping.ModelUrl;
 import com.toy.getyourfriday.service.UpdateService;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
