+++ src/main/java/com/toy/getyourfriday/domain/scraping/ModelUrl.java
@@ -1,4 +1,4 @@
-package com.toy.getyourfriday.domain;
+package com.toy.getyourfriday.domain.scraping;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
