+++ src/test/java/com/toy/getyourfriday/service/ScrapingManagerTest.java
@@ -3,7 +3,7 @@ package com.toy.getyourfriday.service;
 import com.toy.getyourfriday.component.ModelUrlParser;
 import com.toy.getyourfriday.component.ProductContainer;
 import com.toy.getyourfriday.config.ApplicationConfig;
-import com.toy.getyourfriday.domain.ModelUrl;
+import com.toy.getyourfriday.domain.scraping.ModelUrl;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
