+++ src/test/java/com/toy/getyourfriday/service/ScrapingTaskTest.java
@@ -3,7 +3,7 @@ package com.toy.getyourfriday.service;
 import com.toy.getyourfriday.component.ModelUrlParser;
 import com.toy.getyourfriday.component.ProductContainer;
 import com.toy.getyourfriday.config.ApplicationConfig;
-import com.toy.getyourfriday.domain.WebScraper;
+import com.toy.getyourfriday.domain.scraping.WebScraper;
 import com.toy.getyourfriday.service.ScrapingManager.ScrapingTask;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
